gen_wordcloud/pyproject.toml
[project]
name = "gen_wordcloud"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fire>=0.7.0",
    "japanize-matplotlib>=1.1.3",
 "wordcloud>=1.9.4",
]
